refactor(server): file structure
