+++ server/src/api/api.editor.ts
@@ -3,7 +3,7 @@ import {
     ApiSegment,
     EHttpState
 }                        from 'server-modules';
-import { getCollection } from '../database/database';
+import { getCollection } from '../database';
 
 const randomId = (length: number = 16): string => {
     let str = '';
